utils folder renamed as it was leading to a false warning
